pyproject.toml
[tool.poetry]
name = "cursor-agents"
version = "0.1.0"
description = ""
authors = ["Drjunchenfeng <fengjunchen@sparkedu.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.68.2"
mcp = "^1.5.0"
httpx = "^0.28.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
